changes Products name to ProductsList
